Refactor: category 도메인 Service 전환 (#214)
* [#213] Refactor: Category Executor 패턴을 단일 CategoryService로 통합


* [#213] Test: Category Executor 테스트를 CategoryServiceIntegrationTest로 통합


* [#213] Chore: CodeRabbit 코드 리뷰 기준 YAML 설정 추가


* [#213] Docs: CLAUDE.md 아키텍처 기술 및 테스트 가이드 최신화


* [#213] Fix: DataIntegrityViolationException cause 검사 및 YAML 이벤트 리스너 규칙 명확화


* [#213] Test: 캐시 테스트 Cache null 체크 누락 보완


---------
