+++ CLAUDE.md
@@ -35,33 +35,30 @@ controller/ / dto/{command,request,response}/ / entity/ / repository/ / service/
 ### 의존성 방향
 
 ```
-Controller → Service → Executor(Finder/Manager/...) → Repository
+Controller → Service → Repository
+Controller → Facade → Service → Repository
 ```
 
-- Controller는 Service만 호출 (Executor 직접 호출 금지)
-- Service는 Repository 직접 호출 금지, 항상 Executor를 통해 접근
+- Controller는 Service 또는 Facade만 호출
+- Facade는 2개 이상 도메인의 상태 변경을 오케스트레이션할 때만 생성
+- Service는 자기 도메인 Repository + 필요 시 다른 도메인 Repository 직접 주입 가능
+- Service가 다른 도메인 Service를 직접 호출하는 것은 금지 (크로스 도메인 상태 변경은 Facade 담당)
 - 역방향 의존성 절대 금지
 
-### Executor 패턴
+### 마이그레이션 중인 도메인 (Executor → Service+Facade)
 
-Service는 역할별 컴포넌트에 위임한다. 기본적으로 `*Manager`를 사용하며, 특정 역할의 메서드가 2개 이상이 되면 해당 역할의 컴포넌트로 분리한다.
+food/refrigerator/member/notification 도메인은 현재 기존 Executor 패턴(`*Finder`, `*Manager`, `*Modifier`, `*Remover`, `*Appender`, `*Linker`, `*Generator`) 클래스가 남아 있다.
+마이그레이션 진행 중이므로 기존 Executor 클래스를 수정하는 작업은 허용하되, 새 Executor 클래스를 추가하는 것은 금지한다.
+마이그레이션 순서 및 전략은 [`docs/migration-service-facade.md`](docs/migration-service-facade.md) 참조.
 
-| 컴포넌트 | 역할 |
-|----------|------|
-| `*Manager` | 기본 컴포넌트 |
-| `*Finder` | 조회 메서드가 2개 이상일 때 분리 |
-| `*Modifier` | 수정 메서드가 2개 이상일 때 분리 |
-| `*Appender` | 추가 메서드가 2개 이상일 때 분리 |
-| `*Remover` | 삭제 메서드가 2개 이상일 때 분리 |
-| `*Linker` | 연관관계 메서드가 2개 이상일 때 분리 |
-| `*Generator` | 생성 알고리즘 메서드가 2개 이상일 때 분리 |
-| `*Validator` | 검증 로직이 2개 이상일 때 분리 |
+**완료**: category (CategoryService 단일 서비스로 통합)
+**진행 중**: food, refrigerator, member, notification
 
 ## 핵심 규칙
 
 > 상세 내용은 [`docs/core-rules.md`](docs/core-rules.md) 참조
 
-- **트랜잭션**: Service/Executor에 선언, `Finder`는 `readOnly=true`, Controller 선언 금지
+- **트랜잭션**: Service에 선언, 조회 메서드는 `readOnly=true`, Controller 선언 금지
 - **예외**: `CoreException(ErrorType.XXX)` 사용, `IllegalArgumentException` 금지
 - **엔티티**: `BaseEntity` 상속, 정적 팩토리 `register()`/`create()`, 소프트 딜리트(`entity.delete()`), 삭제는 멱등성 보장
 - **DTO**: 불변 `record`, `request.toCommand()` / `Response.of(entity)` 팩토리 통일
@@ -79,3 +76,4 @@ Service는 역할별 컴포넌트에 위임한다. 기본적으로 `*Manager`를
 | `docs/testing-guide.md` | 테스트 작성 패턴 및 FixtureMonkey 활용 |
 | `docs/core-rules.md` | 트랜잭션·예외·엔티티·DTO·이벤트·로깅 상세 규칙 |
 | `docs/testing-rules.md` | 테스트 원칙 및 Fixture 규칙 |
+| `docs/migration-service-facade.md` | Service + Facade 마이그레이션 전략 및 순서 |
