+++ docs/testing-rules.md
@@ -1,5 +1,12 @@
 # 테스트 작성 규칙
 
+## 선택적 작성 원칙
+
+- 모든 코드에 테스트를 작성하지 않는다. 버그 가능성이 높고 변경이 잦은 핵심 기능에 집중한다.
+- 수명이 짧은 이벤트성 기능은 수동 테스트로 대체할 수 있다.
+- 커버리지 수치가 목표가 아니라, 시스템 신뢰성과 문서화 역할을 하는 테스트가 목표다.
+- 상세 전략은 [`docs/testing-guide.md`](testing-guide.md) 참조
+
 ## 테스트 메서드명
 
 - 명사 나열이 아닌 문장으로 작성한다 (A이면 B이다 / A이면 B가 아니고 C다)
